input-raw: Support devices without button capabilities
Certain multi-touch drivers do not report button press events.
Instead, they rely on two subsequent SYN events without any
touch data changes in between to denote that a touch is no
longer occurring.

Change using_syn from a boolean value to a bit flag.  Record in
it whether the SYN can be used to track releases, which will
be the case if the device is capable of SYN events but not BTN
events.  On any change in measured data (X, Y, or pressure),
flag that the touch is still occurring.  When no data changes
between two SYN events, assume a release has occurred.

Combine the multi-touch and single-touch absolute coordinate
cases to reduce duplicate code.